# shared/api/bundles/subscription.openapi.v1.bundle.yaml
openapi: 3.0.3
info:
  title: SubscriptionService API
  version: 1.0.0
  description: API для работы с данными о подписке
tags:
  - name: Subscription
    description: Операции с данными о подписке — получение и обновление информации
paths:
  /api/v1/subscriptions:
    post:
      summary: Create subscription
      operationId: CreateSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_subscription_request'
      tags:
        - Subscription
      responses:
        '201':
          description: Subscription successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/subscription'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '404':
          description: Subscription data for specified city not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/subscriptions/{uuid}:
    parameters:
      - $ref: '#/components/parameters/uuid'
    get:
      summary: Get subscription data for a uuid
      operationId: GetSubscriptionByUuid
      tags:
        - Subscription
      responses:
        '200':
          description: Subscription information successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/subscription'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '404':
          description: Subscription data for specified uuid not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/subscriptions/{uuid}/pay:
    parameters:
      - $ref: '#/components/parameters/uuid'
    post:
      summary: Pay subscription
      operationId: PaySubscriptionByUuid
      tags:
        - Subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pay_subscription_request'
      responses:
        '200':
          description: Subscription paid successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pay_subscription_response'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '409':
          description: Subscription already paid or cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/conflict_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/subscriptions/{uuid}/cancel:
    parameters:
      - $ref: '#/components/parameters/uuid'
    post:
      summary: Cancel subscription
      operationId: CancelSubscriptionByUuid
      tags:
        - Subscription
      responses:
        '204':
          description: Subscription cancelled successfully
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '409':
          description: Subscription already cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/conflict_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
components:
  schemas:
    payment_method:
      type: string
      enum:
        - CARD
        - SBP
        - WALLET
    create_subscription_request:
      type: object
      required:
        - plan_id
        - payment_method
      properties:
        plan_id:
          type: string
          format: uuid
        payment_method:
          $ref: '#/components/schemas/payment_method'
    generic_error:
      type: object
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
        message:
          type: string
          description: Описание ошибки
    subscription_status:
      type: string
      enum:
        - PENDING
        - AWAITING_PAYMENT
        - ACTIVE
        - EXPIRED
        - CANCELLED
    subscription:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/subscription_status'
        plan_id:
          type: string
        plan_name:
          type: string
        amount:
          type: integer
          format: int64
          description: Стоимость в копейках, берётся из тарифа
        currency:
          type: string
          description: Валюта, берётся из тарифа
        payment_method:
          $ref: '#/components/schemas/payment_method'
        expires_at:
          type: string
          format: date-time
          description: Заполняется после оплаты
    bad_request_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 400
        message:
          type: string
          description: Описание ошибки
          example: 'Bad Request: Invalid parameter format'
    not_found_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 404
        message:
          type: string
          description: Описание ошибки
          example: Subscription for uuid 'Unknown' not found
    internal_server_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 500
        message:
          type: string
          description: Описание ошибки
          example: Internal server error occurred
    pay_subscription_request:
      type: object
      required:
        - payment_method
      properties:
        payment_method:
          $ref: '#/components/schemas/payment_method'
    pay_subscription_response:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/subscription_status'
        payment_id:
          type: string
          format: uuid
        payment_method:
          $ref: '#/components/schemas/payment_method'
        expires_at:
          type: string
          format: date-time
    conflict_error:
      type: object
      required:
        - message
      properties:
        code:
          type: string
          description: HTTP-код ошибки
        message:
          type: string
          description: Описание ошибки
  parameters:
    uuid:
      name: uuid
      in: path
      required: true
      description: Идентификатор подписки, для которой запрашиваются данные о подписке
      schema:
        type: string
        minLength: 1
        maxLength: 100
        example: 0000-0000-0000-0000
x-ogen:
  target: ./pkg/openapi/subscription/v1
  package: subscription_v1
  clean: true